+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All significant modifications to this project will be recorded in this file. The
 
 - Removed goal index from the `status <questName> INACTIVE` command
 - Improved the `generateSaveFile()` function to ensure that a subquest's `status` command always after its parent quest's `status` command
+- Improved logical consistency of the `onNewQuestStatus` message
 
 ## [0.1.0] - 2024-05-31